move variant shift to binf module
